requirements.txt
aiohttp==3.7.4
pytest==6.2.5
pytest-asyncio==0.16.0
aioresponses==0.7.2
pytest-aiohttp==0.3.0
marshmallow-dataclass==8.5.3
aiobotocore==1.4.2
python-dotenv
